Implement new AssetGraph and simply create it by converting old AssetGraph
